Public

Public Documentation

Documentation for BioSymbols.jl's public interface.

See Internal Documentation for internal package docs.

Contents

Public Documentation

Contents

Index

Public Interface

Index

BioSymbols.AA_A

BioSymbols.AA_B

BioSymbols.AA_C

BioSymbols.AA_D

BioSymbols.AA_E

BioSymbols.AA_F

BioSymbols.AA_G

BioSymbols.AA_Gap

BioSymbols.AA_H

BioSymbols.AA_I

BioSymbols.AA_J

BioSymbols.AA_K

BioSymbols.AA_L

BioSymbols.AA_M

BioSymbols.AA_N

BioSymbols.AA_O

BioSymbols.AA_P

BioSymbols.AA_Q

BioSymbols.AA_R

BioSymbols.AA_S

BioSymbols.AA_T

BioSymbols.AA_Term

BioSymbols.AA_U

BioSymbols.AA_V

BioSymbols.AA_W

BioSymbols.AA_X

BioSymbols.AA_Y

BioSymbols.AA_Z

BioSymbols.ACGT

BioSymbols.ACGTN

BioSymbols.ACGU

BioSymbols.ACGUN

BioSymbols.DNA_A

BioSymbols.DNA_B

BioSymbols.DNA_C

BioSymbols.DNA_D

BioSymbols.DNA_G

BioSymbols.DNA_Gap

BioSymbols.DNA_H

BioSymbols.DNA_K

BioSymbols.DNA_M

BioSymbols.DNA_N

BioSymbols.DNA_R

BioSymbols.DNA_S

BioSymbols.DNA_T

BioSymbols.DNA_V

BioSymbols.DNA_W

BioSymbols.DNA_Y

BioSymbols.RNA_A

BioSymbols.RNA_B

BioSymbols.RNA_C

BioSymbols.RNA_D

BioSymbols.RNA_G

BioSymbols.RNA_Gap

BioSymbols.RNA_H

BioSymbols.RNA_K

BioSymbols.RNA_M

BioSymbols.RNA_N

BioSymbols.RNA_R

BioSymbols.RNA_S

BioSymbols.RNA_U

BioSymbols.RNA_V

BioSymbols.RNA_W

BioSymbols.RNA_Y

BioSymbols.AminoAcid

BioSymbols.BioSymbol

BioSymbols.DNA

BioSymbols.NucleicAcid

BioSymbols.RNA

BioSymbols.alphabet

BioSymbols.alphabet

BioSymbols.alphabet

BioSymbols.compatbits

BioSymbols.compatbits

BioSymbols.complement

BioSymbols.gap

BioSymbols.gap

BioSymbols.gap

BioSymbols.isGC

BioSymbols.isambiguous

BioSymbols.isambiguous

BioSymbols.iscertain

BioSymbols.iscertain

BioSymbols.iscompatible

BioSymbols.isgap

BioSymbols.ispurine

BioSymbols.ispyrimidine

Public Interface

BioSymbols.AA_A — Constant.

Alanine

source

BioSymbols.AA_B — Constant.

Aspartic Acid or Asparagine

source

BioSymbols.AA_C — Constant.

Cysteine

source

BioSymbols.AA_D — Constant.

Aspartic Acid

source

BioSymbols.AA_E — Constant.

Glutamic Acid

source

BioSymbols.AA_F — Constant.

Phenylalanine

source

BioSymbols.AA_G — Constant.

Glycine

source

BioSymbols.AA_Gap — Constant.

Amino Acid Gap

source

BioSymbols.AA_H — Constant.

Histidine

source

BioSymbols.AA_I — Constant.

Isoleucine

source

BioSymbols.AA_J — Constant.

Leucine or Isoleucine

source

BioSymbols.AA_K — Constant.

Lysine

source

BioSymbols.AA_L — Constant.

Leucine

source

BioSymbols.AA_M — Constant.

Methionine

source

BioSymbols.AA_N — Constant.

Asparagine

source

BioSymbols.AA_O — Constant.

Pyrrolysine

source

BioSymbols.AA_P — Constant.

Proline

source

BioSymbols.AA_Q — Constant.

Glutamine

source

BioSymbols.AA_R — Constant.

Arginine

source

BioSymbols.AA_S — Constant.

Serine

source

BioSymbols.AA_T — Constant.

Threonine

source

BioSymbols.AA_Term — Constant.

Terminal

source

BioSymbols.AA_U — Constant.

Selenocysteine

source

BioSymbols.AA_V — Constant.

Valine

source

BioSymbols.AA_W — Constant.

Tryptophan

source

BioSymbols.AA_X — Constant.

Unspecified or Unknown Amino Acid

source

BioSymbols.AA_Y — Constant.

Tyrosine

source

BioSymbols.AA_Z — Constant.

Glutamine or Glutamic Acid

source

BioSymbols.ACGT — Constant.

ACGT

Unambiguous DNA.

Examples

julia> ACGT
(DNA_A, DNA_C, DNA_G, DNA_T)

source

BioSymbols.ACGTN — Constant.

ACGTN

Unambiguous DNA and DNA_N.

Examples

julia> ACGTN
(DNA_A, DNA_C, DNA_G, DNA_T, DNA_N)

source

BioSymbols.ACGU — Constant.

ACGU

Unambiguous RNA.

Examples

julia> ACGU
(RNA_A, RNA_C, RNA_G, RNA_U)

source

BioSymbols.ACGUN — Constant.

ACGUN

Unambiguous RNA and RNA_N.

Examples

julia> ACGUN
(RNA_A, RNA_C, RNA_G, RNA_U, RNA_N)

source

BioSymbols.DNA_A — Constant.

DNA Adenine

source

BioSymbols.DNA_B — Constant.

DNA Cytosine, Guanine or Thymine

source

BioSymbols.DNA_C — Constant.

DNA Cytosine

source

BioSymbols.DNA_D — Constant.

DNA Adenine, Guanine or Thymine

source

BioSymbols.DNA_G — Constant.

DNA Guanine

source

BioSymbols.DNA_Gap — Constant.

DNA Gap

source

BioSymbols.DNA_H — Constant.

DNA Adenine, Cytosine or Thymine

source

BioSymbols.DNA_K — Constant.

DNA Guanine or Thymine

source

BioSymbols.DNA_M — Constant.

DNA Adenine or Cytosine

source

BioSymbols.DNA_N — Constant.

DNA Adenine, Cytosine, Guanine or Thymine

source

BioSymbols.DNA_R — Constant.

DNA Adenine or Guanine

source

BioSymbols.DNA_S — Constant.

DNA Cytosine or Guanine

source

BioSymbols.DNA_T — Constant.

DNA Thymine

source

BioSymbols.DNA_V — Constant.

DNA Adenine, Cytosine or Guanine

source

BioSymbols.DNA_W — Constant.

DNA Adenine or Thymine

source

BioSymbols.DNA_Y — Constant.

DNA Cytosine or Thymine

source

BioSymbols.RNA_A — Constant.

RNA Adenine

source

BioSymbols.RNA_B — Constant.

RNA Cytosine, Guanine or Uracil

source

BioSymbols.RNA_C — Constant.

RNA Cytosine

source

BioSymbols.RNA_D — Constant.

RNA Adenine, Guanine or Uracil

source

BioSymbols.RNA_G — Constant.

RNA Guanine

source

BioSymbols.RNA_Gap — Constant.

RNA Gap

source

BioSymbols.RNA_H — Constant.

RNA Adenine, Cytosine or Uracil

source

BioSymbols.RNA_K — Constant.

RNA Guanine or Uracil

source

BioSymbols.RNA_M — Constant.

RNA Adenine or Cytosine

source

BioSymbols.RNA_N — Constant.

RNA Adenine, Cytosine, Guanine or Uracil

source

BioSymbols.RNA_R — Constant.

RNA Adenine or Guanine

source

BioSymbols.RNA_S — Constant.

RNA Cytosine or Guanine

source

BioSymbols.RNA_U — Constant.

RNA Uracil

source

BioSymbols.RNA_V — Constant.

RNA Adenine, Cytosine or Guanine

source

BioSymbols.RNA_W — Constant.

RNA Adenine or Uracil

source

BioSymbols.RNA_Y — Constant.

RNA Cytosine or Uracil

source

BioSymbols.AminoAcid — Type.

An amino acid type.

source

BioSymbols.BioSymbol — Type.

The BioSymbol type is an abstract type that represents any kind of biological symbol that may appear in data such as biological sequences, SNP datasets and more.

Every abstract or concrete subtype of BioSymbol is expected to have the following methods defined:

isterm bytemask prefix type_text

source

BioSymbols.DNA — Type.

A deoxyribonucleic acid type.

source

BioSymbols.NucleicAcid — Type.

An abstract nucleic acid type.

source

BioSymbols.RNA — Type.

A ribonucleic acid type.

source

BioSymbols.alphabet — Method.

alphabet(AminoAcid)

Get all symbols of AminoAcid in sorted order.

Examples

julia> alphabet(AminoAcid)
(AA_A, AA_R, AA_N, AA_D, AA_C, AA_Q, AA_E, AA_G, AA_H, AA_I, AA_L, AA_K, AA_M, AA_F, AA_P, AA_S, AA_T, AA_W, AA_Y, AA_V, AA_O, AA_U, AA_B, AA_J, AA_Z, AA_X, AA_Term, AA_Gap)

julia> issorted(alphabet(AminoAcid))
true

source

BioSymbols.alphabet — Method.

alphabet(DNA)

Get all symbols of DNA in sorted order.

Examples

julia> alphabet(DNA)
(DNA_Gap, DNA_A, DNA_C, DNA_M, DNA_G, DNA_R, DNA_S, DNA_V, DNA_T, DNA_W, DNA_Y, DNA_H, DNA_K, DNA_D, DNA_B, DNA_N)

julia> issorted(alphabet(DNA))
true

source

BioSymbols.alphabet — Method.

alphabet(RNA)

Get all symbols of RNA in sorted order.

Examples

julia> alphabet(RNA)
(RNA_Gap, RNA_A, RNA_C, RNA_M, RNA_G, RNA_R, RNA_S, RNA_V, RNA_U, RNA_W, RNA_Y, RNA_H, RNA_K, RNA_D, RNA_B, RNA_N)

julia> issorted(alphabet(RNA))
true

source

BioSymbols.compatbits — Method.

compatbits(aa::AminoAcid)

Return the compatibility bits of aa as UInt32.

Examples

julia> compatbits(AA_A)
0x00000001

julia> compatbits(AA_J)
0x00000600

source

BioSymbols.compatbits — Method.

compatbits(nt::NucleicAcid)

Return the compatibility bits of nt as UInt8.

Examples

julia> compatbits(DNA_A)
0x01

julia> compatbits(DNA_C)
0x02

julia> compatbits(DNA_N)
0x0f

source

BioSymbols.complement — Method.

complement(nt::NucleicAcid)

Return the complementary nucleotide of nt.

This function returns the union of all possible complementary nucleotides.

Examples

julia> complement(DNA_A)
DNA_T

julia> complement(DNA_N)
DNA_N

julia> complement(RNA_U)
RNA_A

source

BioSymbols.gap — Method.

gap(AminoAcid)

Return AA_Gap.

source

BioSymbols.gap — Method.

gap(DNA)

Return DNA_Gap.

source

BioSymbols.gap — Method.

gap(RNA)

Return RNA_Gap.

source

BioSymbols.isGC — Method.

isGC(nt::NucleicAcid)

Test if nt is surely either guanine or cytosine.

source

BioSymbols.isambiguous — Method.

isambiguous(aa::AminoAcid)

Test if aa is an ambiguous amino acid.

source

BioSymbols.isambiguous — Method.

isambiguous(nt::NucleicAcid)

Test if nt is an ambiguous nucleotide.

source

BioSymbols.iscertain — Method.

iscertain(aa::AminoAcid)

Test if aa is a non-ambiguous amino acid.

source

BioSymbols.iscertain — Method.

iscertain(nt::NucleicAcid)

Test if nt is a non-ambiguous nucleotide e.g. ACGT.

source

BioSymbols.iscompatible — Method.

iscompatible(x::S, y::S) where S <: BioSymbol

Test if x and y are compatible with each other.

Examples

julia> iscompatible(AA_A, AA_R)
false

julia> iscompatible(AA_A, AA_X)
true

julia> iscompatible(DNA_A, DNA_A)
true

julia> iscompatible(DNA_C, DNA_N)  # DNA_N can be DNA_C
true

julia> iscompatible(DNA_C, DNA_R)  # DNA_R (A or G) cannot be DNA_C
false

source

BioSymbols.isgap — Method.

isgap(symbol::BioSymbol)

Test if symbol is a gap.

source

BioSymbols.ispurine — Method.

ispurine(nt::NucleicAcid)

Test if nt is surely a purine.

source

BioSymbols.ispyrimidine — Method.

ispyrimidine(nt::NucleicAcid)

Test if nt is surely a pyrimidine.

source